simplifying global gitignore logic

+++ zsh/.zshrc
@@ -61,13 +61,11 @@ fi
 
 # GIT SETUP
 # set up git global ignore if git is present
-if command -v git > /dev/null; then
-  global_gitignore=$HOME/.config/git/ignore # default loc: https://git-scm.com/docs/gitignore
+global_gitignore=$HOME/.config/git/ignore # default loc: https://git-scm.com/docs/gitignore
+if command -v git > /dev/null && [[ ! -f $global_gitignore ]]; then
   mkdir -p $(dirname $global_gitignore)
-  if [[ ! -f $global_gitignore ]]; then
-    echo $'# Globally Ignored Files\n' > $global_gitignore
-    curl -s https://www.toptal.com/developers/gitignore/api/macos,vim,linux,jetbrains+all >> $global_gitignore
-  fi
+  echo $'# Globally Ignored Files\n' > $global_gitignore
+  curl -s https://www.toptal.com/developers/gitignore/api/macos,vim,linux,jetbrains+all >> $global_gitignore
 
   # set pull method if not already specified
   grep -q 'pull' ~/.gitconfig || git config --global pull.rebase true
